backend/requirements.txt
Django==1.11
django-ckeditor==5.4.0
django-cors-headers==2.1.0
django-js-asset==1.0.0
djangorestframework==3.7.7
mysqlclient==1.3.12
Pillow==5.0.0
python-decouple==3.1
pytz==2017.3
